+++ be/src/vec/data_types/serde/data_type_decimal_serde.cpp
@@ -282,7 +282,7 @@ Status DataTypeDecimalSerDe<T>::write_column_to_orc(const std::string& timezone,
 template <typename T>
 
 Status DataTypeDecimalSerDe<T>::deserialize_column_from_fixed_json(
-        IColumn& column, Slice& slice, int rows, uint64_t* num_deserialized,
+        IColumn& column, Slice& slice, uint64_t rows, uint64_t* num_deserialized,
         const FormatOptions& options) const {
     if (rows < 1) [[unlikely]] {
         return Status::OK();
@@ -299,7 +299,7 @@ Status DataTypeDecimalSerDe<T>::deserialize_column_from_fixed_json(
 
 template <typename T>
 void DataTypeDecimalSerDe<T>::insert_column_last_value_multiple_times(IColumn& column,
-                                                                      int times) const {
+                                                                      uint64_t times) const {
     if (times < 1) [[unlikely]] {
         return;
     }
